+++ src/main/java/domain/models/Room.java
@@ -5,21 +5,21 @@ import domain.models.device.Device;
 import java.util.List;
 
 public class Room {
-    private int RoomId;
+    private long RoomId;
     private String name;
     private List<Device> devices;
 
-    public Room(int roomId, String name, List<Device> devices) {
+    public Room(long roomId, String name, List<Device> devices) {
         this.RoomId = roomId;
         this.name = name;
         this.devices = devices;
     }
 
-    public int getRoomId() {
+    public long getRoomId() {
         return RoomId;
     }
 
-    public void setRoomId(int roomId) {
+    public void setRoomId(long roomId) {
         RoomId = roomId;
     }
